+++ eyes.common.java/src/main/java/com/applitools/eyes/IDownloadListener.java
@@ -1,6 +1,0 @@
-package com.applitools.eyes;
-
-public interface IDownloadListener {
-        void onDownloadComplete(String downloadedString);
-        void onDownloadFailed();
-}