+++ eyes.sdk.core/src/main/java/com/applitools/eyes/MatchWindowTask.java
@@ -3,6 +3,7 @@
  */
 package com.applitools.eyes;
 
+import com.applitools.connectivity.api.Response;
 import com.applitools.eyes.capture.AppOutputProvider;
 import com.applitools.eyes.capture.AppOutputWithScreenshot;
 import com.applitools.eyes.fluent.GetFloatingRegion;
@@ -148,7 +149,8 @@ public class MatchWindowTask {
                 int wait = 500;
                 while (retriesLeft-- > 0) {
                     try {
-                        int statusCode = serverConnector.uploadData(bytes, renderingInfo, targetUrl, contentType, mediaType);
+                        Response response = serverConnector.uploadData(bytes, renderingInfo, targetUrl, contentType, mediaType);
+                        int statusCode = response.getStatusCode();
                         if (statusCode == 200 || statusCode == 201) {
                             logger.verbose("upload " + mediaType + " guid " + uuid + "complete.");
                             return targetUrl;